Enum rust_vim_new::rope::Rope

pub enum Rope {
    Leaf(String),
    Node {
        weight: usize,
        left: Box<Rope>,
        right: Box<Rope>,
    },
}

Enumeration of possible values in a Rope Data Structure. Each node can be either a Leaf Node which can be represented by a String, or an internal Node. In the second case, the Node will have its Leaf Nodes and a weight which represents the length of the left Leaf Node.

Variants

Leaf(String)

The sting in case the rope is a Leaf.

Node

Fields

weight: usize

the size of the left part of the Node.

left: Box<Rope>

The left Leaf/Node of the Node.

right: Box<Rope>

The right Leaf/Node of the Node.

Implementations

impl Rope

pub fn new_leaf(data: &str) -> Rope

Creation of a new Leaf node in the binary tree.

Arguments:

data -> the string contained in the leaf.

Returns:

A new leaf Node in the binary tree.

pub fn new_node(left: Rope, right: Rope) -> Rope

Creation of a new internal Node in the binary tree.

Arguments:

left -> the left part of the internal node.

right -> the left part of the internal node.

Returns:

A new internal Node in the binary tree.

pub fn node_length(&self) -> usize

Calculates the length of the strings containing in a Node. This is a recursive funtion. The basic logic is that is that each Node can be either a leaf Node, which means a String or an internal Node that contains futher nodes or leafs. In the first case, the length of the string is getting returned. In the second case, the function calls itself in order until the right Node results into a leaf Node.

Arguments:

&self -> each time represent the node whose length is calculated.

Returns:

The total length of the Node.

pub fn concatenate_ropes(self, rope_to_concat: Rope) -> Rope

Concatenation of two ropes. In case of a Node concatenating with another, this automatically means that the Node has to be an internal one. So, the process has to be simple. The creation of an internal node must be done where the left rope will be the one calling this function and the right one be the other argument.

Arguments:

*rope_to_concat -> The right rope Node that wants to be concatenated.

Returns:

A new Rope internal Node.

pub fn split_rope(self, index: usize) -> (Rope, Rope)

pub fn get_a_substring(&self, start: usize, end: usize) -> String

pub fn collect_substring(&self, start: usize, end: usize, result: &mut String)